requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.10.6
python-dotenv==1.0.1

# LangChain
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

# Vector Search
faiss-cpu==1.8.0
numpy==1.26.4

# Sentence Transformers (for embeddings)
sentence-transformers==3.3.1

# Data Processing
pandas==2.2.3
plotly==5.24.1

# OpenAI
openai==1.51.0
httpx==0.27.2
tiktoken

# Email
resend

# Streamlit
streamlit>=1.32.0